sql: table structure

+++ sql/updates/world/3.3.5/2017_xx_xx_xx_world.sql
@@ -4,7 +4,7 @@ ALTER TABLE `gossip_menu` CHANGE `text_id` `TextID` mediumint(8) unsigned NOT NU
 
 -- gossip_menu_option
 ALTER TABLE `gossip_menu_option` CHANGE `menu_id` `MenuID` smallint(5) unsigned NOT NULL DEFAULT '0';
-ALTER TABLE `gossip_menu_option` CHANGE `id` `OptionID` mediumint(8) unsigned NOT NULL DEFAULT '0';
+ALTER TABLE `gossip_menu_option` CHANGE `id` `OptionID` smallint(5) unsigned NOT NULL DEFAULT '0';
 ALTER TABLE `gossip_menu_option` CHANGE `option_icon` `OptionIcon` mediumint(8) unsigned NOT NULL DEFAULT '0';
 ALTER TABLE `gossip_menu_option` CHANGE `option_text` `OptionText` text;
 ALTER TABLE `gossip_menu_option` CHANGE `option_id` `OptionType` tinyint(3) unsigned NOT NULL DEFAULT '0';
@@ -18,8 +18,8 @@ ALTER TABLE `gossip_menu_option` CHANGE `box_text` `BoxText` text;
 -- gossip_menu_option_locale
 DROP TABLE IF EXISTS `gossip_menu_option_locale`;
 CREATE TABLE IF NOT EXISTS `gossip_menu_option_locale` (
-  `MenuID` smallint(5) unsigned NOT NULL DEFAULT '0',
-  `OptionID` mediumint(8) unsigned NOT NULL DEFAULT '0',
+  `MenuID` smallint(6) unsigned NOT NULL DEFAULT '0',
+  `OptionID` smallint(6) unsigned NOT NULL DEFAULT '0',
   `Locale` varchar(4) NOT NULL,
   `OptionText` text,
   `BoxText` text,
